Module sniffle::prelude

Re-exports commonly used sniffle types, functions, and macros.

Re-exports

pub use crate::protos;

pub use crate::SniffleError;

Modules

pcap

pcapng

Macros

register_dissector

Adds a dissector to be loaded into the default state of a Session. This should only be used if it makes sense for the dissector to be pre-loaded into every Session instance constructed with Session::new() or Session::default().

Structs

AnyPdu

Device

DeviceInjector

DeviceSniffer

DeviceSnifferConfig

EuiAddress

Ipv4Address

Ipv6Address

LogDumper

MacAddress

Packet

Priority

RawPdu

Session

Enums

ConnectionStatus

Traits

Dump

Pdu

PduExt

Sniff

Transmit

Type Definitions

FileSniffer