Module sniffle::uint

A crate for working with unsigned integers of various bit widths.

This crate works as an alternative to existing bitfield crates. The uint types provided by this crate can be used as fields in a struct to emulate bit fields, but the resulting struct is not packed like with actual bit fields. However, the fields are accessible like a normal struct field, where most bitfield crates provide some other way of getting and setting the individual bitfields, and require macros that generate the actual struct. The pack and unpack macros provide the mechanism for packing and unpacking uints like bitfields.

This crate is a utility for the sniffle crate, but can be used outside of sniffle. In sniffle this crate is used for parsing and writing packet headers that have fields with unusual bit widths (i.e. not 8, 16, 32, or 64 bits).

The uints provided by this crate are represented using a builtin unsigned integer (u8, u16, u32, or u64) capable of holding the required number of bits. However, they provide an interface that guarantees a valid n-bit value at all times.

Packing and Unpacking

Packing and unpacking are the bitwise concatenation and spliting of uints, respectively. While these are simple bitwise operations, the pack and unpack macros provide type safety, such that the concatenation of two uints results in a uint whose bit width equals the sum of the bit widths of the unpacked types. For example, packing a U17 and a U23 together results in a U40. This also means that the result of a pack has no padding bits, which could be confusing when trying to determine if padding is in the most or least significant bits of the packed value.

Out-of-the-box, packing can be performed with all tuples of uints where the packed type would be another available uint type. Tuples are packed such that the first (index 0) uint in the tuple is packed into the most significant bits, and the last uint is packed into the least significant bits. This “big endian” order was chosen to logically match with the idea of concatenating the values of the tuple in order.

u128 Feature

The optional u128 feature enables using u128 in order to provide uints U65 to U127. However, the compilation time increases exponentially with more uint types. 64 bits is generally enough for most use cases, so this feature is not enabled by default, but when enabling this feature, be aware that compile times will be very long. Times will vary, but expect a compile time on the order of 15 minutes with this feature enabled.

Example

use sniffle_uint::{self as uint, BitPack, pack, unpack};

// NOTE: The size of BitFields is greater than 1 byte, even if the
//       the bits add up to 8. Each field in this struct is
//       represented with a u8, since each field is less than 8
//       bits and u8 is the smallest builtin uint.
struct BitFields {
    pub a: uint::U3,
    pub b: uint::U4,
    pub c: uint::U1,
}

// NOTE: Perhaps a derive macro will be provided for this pattern
//       in the future.
impl BitPack for BitFields {
    type Packed = u8;

    fn pack(self) -> Self::Packed {
        uint::pack!(self.a, self.b, self.c)
    }

    fn unpack(packed: Self::Packed) -> Self {
        let (a, b, c) = uint::unpack!(packed);
        Self { a, b, c }
    }
}

impl From<u8> for BitFields {
    fn from(value: u8) -> Self {
        Self::unpack(value)
    }
}

impl From<BitFields> for u8 {
    fn from(fields: BitFields) -> Self {
        fields.pack()
    }
}

Macros

pack

Utility to simplify packing uints.

unpack

Utility to simplify unpacking uints.

Structs

U1

A 1-bit unsigned integer.

U2

A 2-bit unsigned integer.

U3

A 3-bit unsigned integer.

U4

A 4-bit unsigned integer.

U5

A 5-bit unsigned integer.

U6

A 6-bit unsigned integer.

U7

A 7-bit unsigned integer.

U9

A 9-bit unsigned integer.

U10

A 10-bit unsigned integer.

U11

A 11-bit unsigned integer.

U12

A 12-bit unsigned integer.

U13

A 13-bit unsigned integer.

U14

A 14-bit unsigned integer.

U15

A 15-bit unsigned integer.

U17

A 17-bit unsigned integer.

U18

A 18-bit unsigned integer.

U19

A 19-bit unsigned integer.

U20

A 20-bit unsigned integer.

U21

A 21-bit unsigned integer.

U22

A 22-bit unsigned integer.

U23

A 23-bit unsigned integer.

U24

A 24-bit unsigned integer.

U25

A 25-bit unsigned integer.

U26

A 26-bit unsigned integer.

U27

A 27-bit unsigned integer.

U28

A 28-bit unsigned integer.

U29

A 29-bit unsigned integer.

U30

A 30-bit unsigned integer.

U31

A 31-bit unsigned integer.

U33

A 33-bit unsigned integer.

U34

A 34-bit unsigned integer.

U35

A 35-bit unsigned integer.

U36

A 36-bit unsigned integer.

U37

A 37-bit unsigned integer.

U38

A 38-bit unsigned integer.

U39

A 39-bit unsigned integer.

U40

A 40-bit unsigned integer.

U41

A 41-bit unsigned integer.

U42

A 42-bit unsigned integer.

U43

A 43-bit unsigned integer.

U44

A 44-bit unsigned integer.

U45

A 45-bit unsigned integer.

U46

A 46-bit unsigned integer.

U47

A 47-bit unsigned integer.

U48

A 48-bit unsigned integer.

U49

A 49-bit unsigned integer.

U50

A 50-bit unsigned integer.

U51

A 51-bit unsigned integer.

U52

A 52-bit unsigned integer.

U53

A 53-bit unsigned integer.

U54

A 54-bit unsigned integer.

U55

A 55-bit unsigned integer.

U56

A 56-bit unsigned integer.

U57

A 57-bit unsigned integer.

U58

A 58-bit unsigned integer.

U59

A 59-bit unsigned integer.

U60

A 60-bit unsigned integer.

U61

A 61-bit unsigned integer.

U62

A 62-bit unsigned integer.

U63

A 63-bit unsigned integer.

Traits

BitPack

A trait that defines how to pack and unpack uints.

FromMasked

IntoMasked

Type Definitions

U8

A 8-bit unsigned integer.

U16

A 16-bit unsigned integer.

U32

A 32-bit unsigned integer.

U64

A 64-bit unsigned integer.